+++ CHANGES.rst
@@ -7,6 +7,9 @@ New Features
 astropy.config
 ^^^^^^^^^^^^^^
 
+- Add new function ``generate_config`` to generate the configuration file and
+  include it in the documentation. [#10148]
+
 astropy.constants
 ^^^^^^^^^^^^^^^^^